--- src/interpreter/types/evaluationResults.ts
@@ -1,12 +1,0 @@
-export type Value = any;
-
-export interface Finished {
-  status: 'finished';
-  value: Value;
-}
-
-export interface Error {
-  status: 'error';
-}
-
-export type Result = Finished | Error;
