+++ src/interpreter/types/evaluator.ts
@@ -1,7 +1,6 @@
-import { type Instr } from './vmInstruction';
+import { type Instr } from './instruction';
 import { type Memory } from '../../memory/memory';
 import { type Stack } from '../../utils/stack';
-import { type Value } from './evaluationResults';
 
 export type EvaluatorMapping = {
   [InstrType in Instr['type']]: (
@@ -14,3 +13,16 @@ export interface EvaluatorState {
   memory: Memory;
   stash: Stack<Value>;
 }
+
+export type Value = any;
+
+export interface Finished {
+  status: 'finished';
+  value: Value;
+}
+
+export interface Error {
+  status: 'error';
+}
+
+export type Result = Finished | Error;
